+++ src/main/java/com/seulseul/seulseul/dto/alarm/AlarmUpdateDto.java
@@ -7,5 +7,5 @@ public class AlarmUpdateDto {
     private Long id;
     private Long base_route_id;
     private Long alarmTime;
-    private Long alarmTerm;
+    private int alarmTerm;
 }
